+++ apps/server/src/features/auth/handlers/create-registration-token-handler.ts
@@ -1,8 +1,10 @@
 import { db } from '../../../db-config';
 import { createHandler } from '../../../utils/create-handler';
+import { getDomainUrl } from '../../../utils/get-domain-url';
 import { createJWT } from '../../../utils/jwt';
 import { loadEnvVariable } from '../../../utils/load-env-variable';
 import { sendEmail } from '../../../utils/send-email';
+import { createEmailHTML } from '../util/create-email-html';
 
 export const createRegistrationTokenHandler = createHandler(async (req, res) => {
   const { email } = req.data;
@@ -19,17 +21,12 @@ export const createRegistrationTokenHandler = createHandler(async (req, res) => 
     },
   );
 
-  const domainUrl = loadEnvVariable('DOMAIN_URL', false) || 'http://localhost:5173';
-  const html = `
-      <html>
-        <h1>Tervehdys!</h1>
-        <p>
-          Olet pyytänyt rekisteröityä <strong>e-markan</strong> käyttäjäksi. Jos tämä ei ollut sinä, voit jättää tämän viestin huomiotta. <br/>
-          Muussa tapauksessa ole hyvä ja napauta <a href="${domainUrl}/#/register?token=${token}">tätä linkkiä</a> jatkaaksesi rekisteröitymisprosessia.
-        </p>
-        <i>Ystävällisin terveisin, e-Markka tiimi.</i>
-      </html>
-    `;
+  const domainUrl = getDomainUrl();
+  const html = createEmailHTML({
+    title: 'Tervehdys!',
+    bodyText: `Olet pyytänyt rekisteröityä <strong>e-markan</strong> käyttäjäksi. Jos tämä ei ollut sinä, voit jättää tämän viestin huomiotta. <br/>
+          Muussa tapauksessa ole hyvä ja napauta <a href="${domainUrl}/register?token=${token}">tätä linkkiä</a> jatkaaksesi rekisteröitymisprosessia.`,
+  });
 
   const emailRes = await sendEmail({
     subject: 'Rekisteröityminen e-Markan käyttäjäksi',