+++ apps/server/src/types/session.ts
@@ -1,4 +1,5 @@
 export type TSession = {
+  token: string;
   user: {
     id: string;
     email: string;