+++ apps/server/src/utils/get-domain-url.ts
@@ -1,0 +1,5 @@
+import { loadEnvVariable } from './load-env-variable';
+
+export function getDomainUrl() {
+  return loadEnvVariable('DOMAIN_URL', false) || 'http://localhost:5173';
+}