+++ packages/schemas/index.ts
@@ -4,3 +4,4 @@ export * from './src/features/auth/register-user-schema.js';
 export * from './src/features/auth/verify-email-schema.js';
 export * from './src/features/transactions/transaction-schema.js';
 export * from './src/features/auth/email-schema.js';
+export * from './src/features/auth/reset-password-schema.js';